Auther_M2/Auther_M/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
pymysql==1.1.1
mysql-connector-python==8.4.0

# Auth
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Forms & Validation
python-multipart==0.0.9
pydantic==2.8.2
email-validator

# Templates
jinja2

# Twilio
twilio

# Env Config
python-dotenv==1.0.1

# Migration
alembic==1.13.3
